% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/potential_candidates.R
\name{CalcReplaceF}
\alias{CalcReplaceF}
\title{Calculate replacement fishing mortality}
\usage{
CalcReplaceF(stock, rpB, rpR)
}
\arguments{
\item{stock}{XXX}

\item{rpB}{XXX}

\item{rpR}{XXX}
}
\description{
XXX
}

