% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{flindices_to_rbya}
\alias{flindices_to_rbya}
\title{Converts FLIndices object to rbya}
\usage{
flindices_to_rbya(x)
}
\arguments{
\item{x}{An FLIndices object}
}
\description{
The rbya (results by year and age) is a long \code{data.frame}
popular in Reykjavik
}
