% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flr.R
\name{flr_rbya}
\alias{flr_rbya}
\title{xxx}
\usage{
flr_rbya(fl, scale = 1, stk, des, wide = TRUE)
}
\arguments{
\item{fl}{An FLStock object}

\item{scale}{A numeric, specifying re-scaling (default is 1). Not used (yet)}

\item{stk}{Name of the run, if missing (default) will use the directory name
or stock name if FL-object.}

\item{wide}{A boolean (default TRUE). If FALSE values returned as a single
column (named "val"), variable names specified in column named "var"}

\item{dec}{desc if an FL-object}
}
\value{
A tibble containing key metrics by year and age
}
\description{
xxx
}
