% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{flstock_to_rbya}
\alias{flstock_to_rbya}
\title{Converts FLStock object to rbya}
\usage{
flstock_to_rbya(x, scale = 1, project = TRUE, plusgroup = TRUE)
}
\arguments{
\item{x}{An FLStock object}

\item{scale}{scaler used on abundance values (stock in numbers, catches etc)}

\item{project}{A boolean, if TRUE (default), propagates terminal stock numbers
forward by one year (into the assessment year). Note that the weights in the
assessment year are the same as in the terminal year.}

\item{plusgroup}{A boolean, if TRUE (default), last age group is a plus group.
Only used if project is TRUE.}
}
\description{
The rbya (results by year and age) is a long \code{data.frame}
popular in Reykjavik
}

