% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishvice.R
\name{fv_opr}
\alias{fv_opr}
\title{Observed, predicted and residuals}
\usage{
fv_opr(o, scale = 1)
}
\arguments{
\item{o}{An object of class sam, FLSAM or a directory path (for muppet)}

\item{scale}{A scale for the scaleable variables}
}
\value{
A tibble containing key metric by year
}
\description{
Observed, predicted and residuals
}
