% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishvice.R
\name{fv_rbya}
\alias{fv_rbya}
\title{Results by year and age}
\usage{
fv_rbya(o)
}
\arguments{
\item{o}{An object of class FLSAM or FLStock}
}
\value{
A tibble containing key metric by year
}
\description{
Results by year and age
}
