% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.R
\name{get_ices_directory}
\alias{get_ices_directory}
\title{Get a copy of ICES sharepoint directory}
\usage{
get_ices_directory(user_name, password, directory)
}
\arguments{
\item{user_name}{Only the name after the ices-backslash}

\item{password}{Your password}

\item{directory}{The path to the directory you want. I}
}
\description{
Makes a full copy of a sharepoint directory onto your computer.
In theory you can request a full mirror of the whole sharepoint :-)
So use at your own discretion!

Still only function for some directories
}

