% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sam.R
\name{get_sam_directory}
\alias{get_sam_directory}
\title{Get a sam directory from stockassessment.org}
\usage{
get_sam_directory(assessment, user = "user3")
}
\arguments{
\item{assessment}{Name of the assessment}

\item{user}{Name of the user}
}
\description{
The function copies the whole directory of an assessment run from
stockassessment.org to a local directory
}

