% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standard_graphs.R
\name{get_stock_list}
\alias{get_stock_list}
\title{returns the list of all the fishstocks found in the database.}
\usage{
get_stock_list(year)
}
\arguments{
\item{year}{The year of the assessment. If the year is 0 then it will return
all the assessment years. If there are no matches, an empty result set will
be returned.}
}
\value{
Returns the year of the assessment, the species , the fishstock nam,
the ecoregion and the key of the fishstocks/assessment year.
}
\description{
The function is a
wrapper around the ICES webservice API
\href{http://standardgraphs.ices.dk/standardgraphswebservices.asmx/getListStocks}{getListStocks}
}

