% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standard_graphs.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Returns the Summary table of the stock}
\usage{
get_summary_table(key)
}
\arguments{
\item{key}{The key of the assessment. Example would be 1786 for the Cod in
Subarea IV (North Sea), Divison VIId (Eastern Channel) and IIIa West (Skagerrak) (cod-347d)
in the assessment year of 2013. If there are no matches, an empty result set will be returned.}
}
\value{
Returns the year of the assessment, the species , the fishstock nam,
the ecoregion and the key of the fishstocks/assessment year.
}
\description{
The function is a
wrapper around the ICES webservice API
\href{http://standardgraphs.ices.dk/standardgraphswebservices.asmx/getListStocks}{getListStocks}
}

