% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoRMS.R
\name{get_worms_id}
\alias{get_worms_id}
\title{Get the WoRMS id for a taxonomic name}
\usage{
get_worms_id(scientificname = "Gadus morhua")
}
\arguments{
\item{scientificname}{Linnean type of name}
}
\value{
Returns a numerical value (AphiaID):
\itemize{
  \item{NULL: when no matches are found}
  \item{-999: when multiple matches was found}
  \item{an integer: AphiaID when one exact match was found}
}
}
\description{
Connection to WoRMS webservice \href{http://marinespecies.org/aphia.php?p=soap}{getAphiaID}.
}

