% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_admb.R
\name{getlineswin}
\alias{getlineswin}
\title{Gets certain lines from file}
\usage{
getlineswin(file, line1, line2)
}
\arguments{
\item{file}{Name of the file}

\item{line1}{Integer, first line}

\item{line2}{Integer, last line}
}
\value{
A xxx
}
\description{
Some longer text here
}
