% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/potential_candidates.R
\name{gg_crayola}
\alias{gg_crayola}
\title{Crayola plot}
\usage{
gg_crayola(data, x, y, fill, ...)
}
\arguments{
\item{data}{XXX}

\item{x}{XXX}

\item{y}{XXX}

\item{fill}{XXX}

\item{...}{XXX}
}
\description{
Crayola plot
}

