% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.R
\name{ices_standard_graph}
\alias{ices_standard_graph}
\title{ICES standard graph}
\usage{
ices_standard_graph(rby, graph_name)
}
\arguments{
\item{rby}{data.frame containing ...}

\item{graph_name}{Name of the graph to be created}
}
\description{
XXX
}
