% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ices.R
\name{ices_standard_graph_cod}
\alias{ices_standard_graph_cod}
\title{ICES standard graph for Icelandic cod}
\usage{
ices_standard_graph_cod(rby, graph_name)
}
\arguments{
\item{rby}{data.frame containing ...}

\item{graph_name}{Name of the graph to be created}
}
\description{
The intent is to make this graph generic, now only reasonable
for iCod.
}

