% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kvoti.R
\name{kvota_stada}
\alias{kvota_stada}
\title{Upplýsingar um kvótastöðu}
\usage{
kvota_stada(sid = c(1, 2, 3, 5, 6, 7, 8, 9, 11, 22, 23, 24, 25, 27, 28, 30,
  31, 34, 36, 40, 41, 61))
}
\arguments{
\item{sid}{Tegundarnúmer}
}
\description{
XXX
}

