% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lexicon}
\alias{lexicon}
\title{A lookup of variables between rbx, flr, ...}
\format{
A data frame with 13 rows and 2 variables:
\describe{
  \item{rvk}{variable names used in rvk}
  \item{slot}{variable names used in flr}
  ...
}
}
\usage{
lexicon
}
\description{
A lookup of variables between rbx, flr, ...
}
\keyword{datasets}
