% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marinetraffic.R
\name{mt_vesselinfo}
\alias{mt_vesselinfo}
\title{Vessel info from marinetraffic.com}
\usage{
mt_vesselinfo(mtid = 294305)
}
\arguments{
\item{mtid}{Marine traffic vessel id}
}
\value{
A tibble
}
\description{
Vessel info from marinetraffic.com
}
