% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muppet_mcmc.R
\name{mup_mcmc}
\alias{mup_mcmc}
\title{read mcmc file}
\usage{
mup_mcmc(path, file, std = TRUE)
}
\arguments{
\item{path}{The directory containing the mcmc files}

\item{file}{A character vector containing the file name. If missing (DEFAULT)
reads in all year-results}

\item{std}{A boolean, if TRUE (default) returns niceer var names}
}
\value{
a tibble three variables: year, val, var and iter
}
\description{
read mcmc file
}
