% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muppet.R
\name{mup_rbx}
\alias{mup_rbx}
\title{muppet_rbx}
\usage{
mup_rbx(path, scale = 1, assyear, run, wide = TRUE)
}
\arguments{
\item{path}{A directory path}

\item{scale}{A scaler (default 1)}

\item{assyear}{Assessment year, if missing (default) will use the year after the last
catch at age input}

\item{run}{Name of the run, if missing (default) will use the directory name}

\item{wide}{A boolean indicating if returned table wide (default TRUE). If FALSE variable are
return within column 'var' and values in column 'val'.}
}
\value{
A tibble
}
\description{
reads muppet results
}
\examples{
rbx <- mup_rbx(path = system.file("mup/smx", package = "fishvice"), scale = 1000)
names(rbx)
}
