% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dplyr_friendly_functions.R
\name{r2lat}
\alias{r2lat}
\title{Returns the latitude for a statistical square (tilkynningaskildureitur)}
\usage{
r2lat(x)
}
\arguments{
\item{x}{Numerical vector containing statistical square}
}
\description{
A dplyr-friendly variant of the geo::r2d, returning only a
vector of latitudes. To be used inside mutate.
}

