% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dplyr_friendly_functions.R
\name{r2lon}
\alias{r2lon}
\title{Returns the lon for a statistical square (tilkynningaskildureitur)}
\usage{
r2lon(x)
}
\arguments{
\item{x}{Numerical vector containing statistical square}
}
\description{
A dplyr-friendly variant of the geo::r2d, returning only a
vector of longitudes. To be used inside mutate.
}

