% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rba_icod}
\alias{rba_icod}
\title{Results by age}
\format{An object of class \code{data.frame} with 12 rows and 12 columns.}
\usage{
rba_icod
}
\description{
Example file: Cod results by age
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

