% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvk_to_converter.R
\name{rbx_to_flindex}
\alias{rbx_to_flindex}
\title{rbx_to_flindex}
\usage{
rbx_to_flindex(rbx, sur = "U1", time = c(3/12, 3.5/12))
}
\arguments{
\item{rbx}{xxx}

\item{sur}{xxx}

\item{time}{xxx}
}
\value{
FLIndex
}
\description{
rbx_to_flindex
}
