% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rby_icod}
\alias{rby_icod}
\title{Results by year}
\format{\preformatted{'data.frame':	61 obs. of  5 variables:
 $ Year       : int  1955 1956 1957 1958 1959 1960 1961 1962 1963 1964 ...
 $ Recruitment: num  151988 152797 170633 220714 289052 ...
 $ SSB        : num  941982 795988 775735 875551 853637 ...
 $ F          : num  0.29 0.29 0.311 0.354 0.322 ...
 $ Landings   : num  545250 486909 455182 517359 459081 ...
}}
\usage{
rby_icod
}
\description{
Example file: Cod results by year.
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

