% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_admb.R
\name{read_hst}
\alias{read_hst}
\title{Reads ADMB hst files}
\usage{
read_hst(file, dir, txt, startyear, names, negative.allowed = FALSE)
}
\arguments{
\item{file}{Name of the file, normally ends with .hst}

\item{dir}{Name of the \emph{directory} that contains the result.}

\item{txt}{The parameters to extract}

\item{startyear}{Assessment start year}

\item{names}{A character vector of length two replacing the default names
\emph{c('value','prop')}}

\item{negative.allowed}{Flag, default is FALSE}
}
\value{
A list, each component being a dataframe with two columns. The
default names are \emph{c('value','prop')} which can be replace by specifying
\emph{names}.
}
\description{
Reads output of ADMB MCMC report containing estimated distributions. The .hst
report contains information about the MCMC analysis: the sample sizes
(specied with the -mcmc command-line option), the step size scaling
factor, the step sizes, and information about the posterior probability
distribution (e.g., the mean, standard deviation, and lower and upper
bounds). For each simulated parameter, a range of values (with step sizes
reported in the "step sizes" section of the .hst file) and their simulated
posterior probabilities is reported. Plotting the first column
(parameter values) on the x-axis and the second column (simulated
probabilities) on the y-axis can be a convenient way to make a visualization
of the posterior probability distribution.
}
