% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_lowestoft.R
\name{read_lowestoft_survey}
\alias{read_lowestoft_survey}
\title{Read in Lowestoft-format survey data}
\usage{
read_lowestoft_survey(filename, format = "data.frame")
}
\arguments{
\item{filename}{The name of the file to read in}

\item{format}{Character, if "data.frame" (default) returns a data.frame
otherwise a list.}
}
\description{
Reads in survey data in the Lowestoft format.
}
