% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/potential_candidates.R
\name{read_mcmc}
\alias{read_mcmc}
\title{Read mcmc files}
\usage{
read_mcmc(path = Path, run = Run, variable = c("catch", "ssb", "f", "n3",
  "refbio"), years, hrates, nIter, sDir = "msyresults")
}
\arguments{
\item{path}{Path to the root directory}

\item{run}{Name of the run directory}

\item{variable}{Names of variables in the output}

\item{years}{The years to read in}

\item{hrates}{The harvest rates to read}

\item{nIter}{Number of iterations that were run}

\item{sDir}{The subdirectory within the Run directory that stores the data}
}
\description{
Works for saithe
}

