% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbx.R
\name{read_muppet}
\alias{read_muppet}
\title{read muppet output}
\usage{
read_muppet(path, Scale = 1, run, assyear)
}
\arguments{
\item{path}{A vector containing the path to the assessment directory}

\item{Scale}{A numeric, specifying rescaling (default is 1)}

\item{run}{Name of the run, if missing (default) will use the directory name}

\item{assyear}{Assessment year, if missing (default) will use the last
year of the catch at age matrix plus 1.}
}
\value{
A list containing tibbles
}
\description{
read muppet output
}
