% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelude.R
\name{read_prelude}
\alias{read_prelude}
\title{Reads prelude files}
\usage{
read_prelude(file, rownames = FALSE, underscore2dot = FALSE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{rownames}{a logical value (default is FALSE) indicating whether the
first column should be treated as row names.}

\item{underscore2dot}{a logical value (default is FALSE) indicating if
underscores in column names should be replaced with dots. Note the default is
the reverse from that in the function s2pre, the latter being a somewhat odd
default.}
}
\description{
Read data files in
\href{http://www.hafro.is/~gunnar/manual/pdf/080_Forleikur.pdf}{prelude format},
which has 2 line headers, the first containing the column name the second
containing dashse.

Based on script in the geo-package, there called s2pre.
}
\author{
hoski@hafro.is
}
\seealso{
\code{\link{write_prelude}} for writing prelude files
}

