% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sam.R
\name{read_sam}
\alias{read_sam}
\title{read_sam}
\usage{
read_sam(directory = "WBcod_2015_short", from_web = FALSE, user = "user3")
}
\arguments{
\item{directory}{The directory name that contains the sam run.}

\item{from_web}{If FALSE (default) read from local directory. If TRUE read
from www.stockassessment.org}

\item{user}{User name if reading from www.stockassessment.org, guest users can use "user3" (default)}
}
\description{
Gets the input files and sam output files from the www.stockassessment.org

Some important pieces of the code are from Anders Nielsen
}

