% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ices.R, R/standard_xml.R
\name{rvk2icesxml}
\alias{rvk2icesxml}
\title{rvk2icesxml}
\usage{
rvk2icesxml(rby, FishStock, AssessmentYear, RecruitmentAge, FAge, rba,
  UnitsWeigths = "tonnes", UnitsRecruits = "millions",
  TypeLandings = "official", Flim, Fpa, Blim, Bpa, FMSY, MSYBtrigger,
  Fmanagement, Bmanagement, RecruitmentLength, Custom1, Custom2, Custom3,
  Custom4, Custom5, VersionStock, NameSystemProducedFile)

rvk2icesxml(rby, FishStock, AssessmentYear, RecruitmentAge, FAge, rba,
  UnitsWeigths = "tonnes", UnitsRecruits = "millions",
  TypeLandings = "official", Flim, Fpa, Blim, Bpa, FMSY, MSYBtrigger,
  Fmanagement, Bmanagement, RecruitmentLength, Custom1, Custom2, Custom3,
  Custom4, Custom5, VersionStock, NameSystemProducedFile)
}
\arguments{
\item{rby}{A data frame containing stock summary data. Column names
besides the Year column which is required can be any of the following:
\itemize{
 \item \emph{Year} - Year
 \item \emph{Low_Recruitment} - XX
 \item \emph{Recruitment} - XXX
 \item \emph{High_Recruitment} - XXX
 \item \emph{Low_TBiomass} - XXX
 \item \emph{TBiomass} - XXX
 \item \emph{High_TBiomass} - XXX
 \item \emph{Low_SSB} - XXX
 \item \emph{SSB} - XXX
 \item \emph{High_SSB} - XXX
 \item \emph{Catches} - XXX
 \item \emph{Landings} - XXX
 \item \emph{Discards} - XXX
 \item \emph{IBC} - XXX
 \item \emph{Unallocated_Removals} - XXX
 \item \emph{YieldSSB} - XXX
 \item \emph{Low_F} - XXX
 \item \emph{F} - XXX
 \item \emph{High_F} - XXX
 \item \emph{F_Landings} - XXX
 \item \emph{F_Discards} - XXX
 \item \emph{F_IBC} - XXX
 \item \emph{F_Unallocated} - XXX
 \item \emph{SoP} - XXX
 \item \emph{Custom1} - XXX
 \item \emph{Custom2} - XXX
 \item \emph{Custom3} - XXX
 \item \emph{Custom4} - XXX
 \item \emph{Custom5} - XXX
}}

\item{FishStock}{A character vector containing stock "name". e.g. "cod-farp", "cod-iceg".}

\item{AssessmentYear}{An integer representing assessment year}

\item{RecruitmentAge}{An integer representing recruitment age}

\item{FAge}{A character vector containg reference fishing moralites, e.g. "F5-10"}

\item{rba}{\code{data.frame} containing yield per recruit input. If missing it is
ignored. If supplied required column names are:
\itemize{
 \item \emph{Age} - age
 \item \emph{M} - natural mortality
 \item \emph{Mat} - maturity
 \item \emph{PF} - partial fishing mortality before spawning
 \item \emph{PM} - partial natural mortality before spawning
 \item \emph{WeSt} - spawning stock weights
 \item \emph{F} - selection pattern
 \item \emph{WeCa} - catch weights
 \item \emph{Fd} - selection pattern of discards
 \item \emph{WeCad} - discard catch weights
 \item \emph{Fi} - selection pattern of industrial waste
 \item \emph{WeCai} - industrial catch weights
}}

\item{UnitsWeigths}{A character, expect any one of "Kilograms", "Tonnes",
"Thousand tonnes", "Million tonnes"}

\item{UnitsRecruits}{A character, expect any one of "Thousands", "Millions",
"Billions"}

\item{TypeLandings}{A character, e.g. "official", "ices estimates".}

\item{Flim}{A numeric value}

\item{Fpa}{A numeric value}

\item{Blim}{A numeric value}

\item{Bpa}{A numeric value}

\item{FMSY}{A numeric value}

\item{MSYBtrigger}{A numeric value}

\item{Fmanagement}{A numeric value}

\item{Bmanagement}{A numeric value}

\item{RecruitmentLength}{A numeric value}

\item{Custom1}{XXX}

\item{Custom2}{XXX}

\item{Custom3}{XXX}

\item{Custom4}{XXX}

\item{Custom5}{XXX}

\item{VersionStock}{XXX}

\item{NameSystemProducedFile}{XXX}

\item{rby}{A data frame containing stock summary data. Column names
besides the Year column which is required can be any of the following:
\itemize{
 \item \emph{Year} - Year
 \item \emph{Low_Recruitment} - XX
 \item \emph{Recruitment} - XXX
 \item \emph{High_Recruitment} - XXX
 \item \emph{Low_TBiomass} - XXX
 \item \emph{TBiomass} - XXX
 \item \emph{High_TBiomass} - XXX
 \item \emph{Low_SSB} - XXX
 \item \emph{SSB} - XXX
 \item \emph{High_SSB} - XXX
 \item \emph{Catches} - XXX
 \item \emph{Landings} - XXX
 \item \emph{Discards} - XXX
 \item \emph{IBC} - XXX
 \item \emph{Unallocated_Removals} - XXX
 \item \emph{YieldSSB} - XXX
 \item \emph{Low_F} - XXX
 \item \emph{F} - XXX
 \item \emph{High_F} - XXX
 \item \emph{F_Landings} - XXX
 \item \emph{F_Discards} - XXX
 \item \emph{F_IBC} - XXX
 \item \emph{F_Unallocated} - XXX
 \item \emph{SoP} - XXX
 \item \emph{Custom1} - XXX
 \item \emph{Custom2} - XXX
 \item \emph{Custom3} - XXX
 \item \emph{Custom4} - XXX
 \item \emph{Custom5} - XXX
}}

\item{FishStock}{A character vector containing stock "name". e.g. "cod-farp", "cod-iceg".}

\item{AssessmentYear}{An integer representing assessment year}

\item{RecruitmentAge}{An integer representing recruitment age}

\item{FAge}{A character vector containg reference fishing moralites, e.g. "F5-10"}

\item{rba}{\\code{data.frame} containing yield per recruit input. If missing it is
ignored. If supplied required column names are:
\itemize{
 \item \emph{Age} - age
 \item \emph{M} - natural mortality
 \item \emph{Mat} - maturity
 \item \emph{PF} - partial fishing mortality before spawning
 \item \emph{PM} - partial natural mortality before spawning
 \item \emph{WeSt} - spawning stock weights
 \item \emph{F} - selection pattern
 \item \emph{WeCa} - catch weights
 \item \emph{Fd} - selection pattern of discards
 \item \emph{WeCad} - discard catch weights
 \item \emph{Fi} - selection pattern of industrial waste
 \item \emph{WeCai} - industrial catch weights
}}

\item{Flim}{A numeric value}

\item{Fpa}{A numeric value}

\item{Blim}{A numeric value}

\item{Bpa}{A numeric value}

\item{FMSY}{A numeric value}

\item{MSYBtrigger}{A numeric value}

\item{Fmanagement}{A numeric value}

\item{Bmanagement}{A numeric value}

\item{RecruitmentLength}{A numeric value}

\item{UnitsWeigths}{A character, expect any one of "Kilograms", "Tonnes",
"Thousand tonnes", "Million tonnes"}

\item{UnitsRecruits}{A character, expect any one of "Thousands", "Millions",
"Billions"}

\item{TypeLandings}{A character, e.g. "official", "ices estimates".}

\item{Custom1}{XXX}

\item{Custom2}{XXX}

\item{Custom3}{XXX}

\item{Custom4}{XXX}

\item{Custom5}{XXX}

\item{VersionStock}{XXX}

\item{NameSystemProducedFile}{XXX}
}
\description{
Converts data.frame to ices xml format for standard graphs. The
function is a brute force coding and does not have any error checking on the
input data.

Converts data.frame to ices xml format for standard graphs. The
function is a brute force coding and does not have any error checking on the
input data.
}

