% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_get_directory}
\alias{sam_get_directory}
\title{Get a sam directory from stockassessment.org}
\usage{
sam_get_directory(assessment, user = "user3")
}
\arguments{
\item{assessment}{Name of the assessment}

\item{user}{Name of the user}
}
\description{
The function copies the whole directory of an assessment run from
stockassessment.org to a local directory
}

