% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_get_residuals}
\alias{sam_get_residuals}
\title{sam_get_residuals}
\usage{
sam_get_residuals(assessment)
}
\arguments{
\item{assessment}{Name of run on stockassessment.org}
}
\value{
XXXX
}
\description{
sam_get_residuals
}
