% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_get_results}
\alias{sam_get_results}
\title{sam_get_results}
\usage{
sam_get_results(assessment = "WBcod_2015_short", user = "user3")
}
\arguments{
\item{assessment}{The directory name of the assessment as specified on
www.stockassessment.org}

\item{user}{User name, guest users can use "user3" (default)}
}
\description{
Gets the input files and sam output files from the www.stockassessment.org

Some important pieces of the code are from Anders Nielsen
}

