% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_ibya_fromsam}
\alias{sam_ibya_fromsam}
\title{sam_ibya_fromsam}
\usage{
sam_ibya_fromsam(fit, scale = 1)
}
\arguments{
\item{fit}{An object of class sam (often named fit by users)}

\item{scale}{A scaler (default 1)}
}
\value{
A tibble
}
\description{
sam_ibya_fromsam
}
