% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_one_observation_ahead_residuals}
\alias{sam_one_observation_ahead_residuals}
\title{One-observation-ahead residuals}
\usage{
sam_one_observation_ahead_residuals(res)
}
\arguments{
\item{res}{An object returned from {stockassessment::residuals}}
}
\value{
A tibble, the "one-observation-ahead-residuals" is in
variable residual.
}
\description{
Note, this is normally just plotted using plot(res)
}
