% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_opr}
\alias{sam_opr}
\title{Observations, predictions and residuals}
\usage{
sam_opr(fit, lgs = TRUE)
}
\arguments{
\item{fit}{A "sam" object}

\item{lgs}{boolean (default TRUE) indicating if data should return
as logs or as ordinary (then FALSE).}
}
\value{
A list of length 2 containing the following:
\itemize{
   \item data: A tibble with the following variables:
   \itemize{
      \item year:
      \item age:
      \item fleet: Name of the fleets
      \item o: Observed value, default log scale
      \item p: Predicted value, default log scale
      \item r: Residuals (always log scale), difference between observed and predicted.
      }
   \item plot: A list containing a plot of the observed and predicted values for each fleet.
   }
}
\description{
Observations, predictions and residuals
}
