% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_process_residuals}
\alias{sam_process_residuals}
\title{Proces residuals}
\usage{
sam_process_residuals(resp)
}
\arguments{
\item{resp}{An object returned from {stockassessment::procres}}
}
\value{
A tibble
}
\description{
Proces residuals
}
