% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_rby}
\alias{sam_rby}
\title{sam_rby}
\usage{
sam_rby(fit)
}
\arguments{
\item{fit}{XXX}
}
\value{
A tibble containing the following variables:
\itemize{
  \item year:
  \item est: Medium value
  \item low: Lower 2.5% quantile??
  \item high: Upper 2.5% quantile??
  \item variable: Name of variable (catch, recruitment, ssb, tsb and fbar)
}
}
\description{
Get assessment summary data from "sam"-object
}
