% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_rbya}
\alias{sam_rbya}
\title{sam_rbya}
\usage{
sam_rbya(fit, data = TRUE, scale = 1, long = TRUE, run)
}
\arguments{
\item{fit}{A "sam" object}

\item{data}{A boolean (default TRUE) specifying if input data should also
be returned.}

\item{scale}{A scaler (default 1)}

\item{long}{A boolean indicating if returned table wide (default TRUE) variables
as names within column 'var'. Alternative (FALSE) not yet active.}

\item{run}{Name of the run}
}
\value{
A tibble, containing at minimum:
\itemize{
  \item year:
  \item age:
  \item n: stock in numbers
  \item f: fishing mortality
}
}
\description{
Extracts stock in numbers (Nay) and fishing mortality (Fay) at
age from object "sam". If a tibble ibya (e.g. generated by {sam_ibya} is
included in the argument, these estimates get added to that tibble.
}
