% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_addendums.R
\name{str_strip_nondigits}
\alias{str_strip_nondigits}
\title{Strip all non digits from a string}
\usage{
str_strip_nondigits(string, out.numeric = TRUE)
}
\arguments{
\item{string}{A character vector}

\item{out.numeric}{Boolean, if TRUE (default) string coerced to numeric.}
}
\description{
Function strips all alphabetic and "special" characters from a string.
Ergo, only digits 0-9 and the decimal (".") is returned.
}
