% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_addendums.R
\name{str_strip_text}
\alias{str_strip_text}
\title{Strip all text from a string}
\usage{
str_strip_text(string)
}
\arguments{
\item{string}{A character vector}
}
\description{
Function strips all alphabetic characters (a-z and A-Z) from a string
}

