% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_addendums.R
\name{str_trim_commas}
\alias{str_trim_commas}
\title{Trim 'commas' from start and end of string}
\usage{
str_trim_commas(string, side = "both")
}
\arguments{
\item{string}{input character vector}

\item{side}{side on which character string is removed (left, right or both)}
}
\description{
Function is enspired by str_trim function in the stringr package.
}

