% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr_addendums.R
\name{str_trim_tab}
\alias{str_trim_tab}
\title{Trim 'tabs' from start and end of string}
\usage{
str_trim_tab(string, side = "both")
}
\arguments{
\item{string}{input character vector}

\item{side}{side on which character string is removed (left, right or both)}
}
\description{
Function is inspired by str_trim function in the stringr package.
The str_trim function is for trimming whitespace, here tabs are trimmed.
}

