% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelude.R
\name{write_prelude}
\alias{write_prelude}
\title{Writes prelude file}
\usage{
write_prelude(data, file = "R.pre", na.replace = "")
}
\arguments{
\item{data}{the object to be written, a \code{matrix} or \code{data.frame}.}

\item{file}{a character string for the output file name (default is "R.pre").}

\item{na.replace}{A character to replace NA with in the output file ("" by default)}
}
\description{
Writes \code{data.frame} or \code{matrix} objects to a file
in
\href{http://www.hafro.is/~gunnar/manual/pdf/080_Forleikur.pdf}{prelude format}.
}
\note{
If file exist it is simply overwritten without any warning
}
\author{
hoski@hafro.is
}
\seealso{
\code{\link{read_prelude}} for reading prelude files
}

