% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/converter.R
\name{FLStock_to_rbya}
\alias{FLStock_to_rbya}
\title{Converts FLStock object to rbya}
\usage{
FLStock_to_rbya(x, scale = 1)
}
\arguments{
\item{x}{An FLStock object}

\item{scale}{scaler used on abundance values (stock in numbers, catches etc)}
}
\description{
The rbya (results by year and age) is a long \code{data.frame}
popular in Reykjavik
}

