% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishvice.R
\name{fv_par}
\alias{fv_par}
\title{Parameters}
\usage{
fv_par(o, scale = 1, run)
}
\arguments{
\item{o}{An object of class sam, FLSAM or a directory path (for muppet)}

\item{scale}{A scale for the scaleable variables}

\item{run}{Name of the run}
}
\value{
A tibble containing key metrics
}
\description{
Parameters
}
