% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ple4}
\alias{ple4}
\title{ple4}
\format{
An object of class \code{FLStock} of dimension 10 x 58 x 1 x 1 x 1 x 1.
}
\usage{
ple4
}
\description{
Example file: North Sea Place 2015 assessment results
}
\author{
David Miller <david.miller@wur.nl>
}
\keyword{datasets}
