% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ple4}
\alias{ple4}
\title{ple4}
\format{\preformatted{Formal class 'FLStock' [package "FLCore"] with 20 slots
  ..@ catch       :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1, 1:58, 1, 1, 1, 1] 78443 88191 109164 117334 118474 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr "all"
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "tonnes"
  ..@ catch.n     :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 32356 50388 75648 50568 36243 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "thousands"
  ..@ catch.wt    :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0.044 0.11 0.194 0.257 0.349 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "kg"
  ..@ discards    :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1, 1:58, 1, 1, 1, 1] 7880 14837 29864 29793 32490 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr "all"
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "tonnes"
  ..@ discards.n  :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 32356 45596 9220 909 961 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "thousands"
  ..@ discards.wt :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0.044 0.104 0.146 0.181 0.206 0.244 0.244 0.231 0 0 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "kg"
  ..@ landings    :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1, 1:58, 1, 1, 1, 1] 70563 73354 79300 87541 85984 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr "all"
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "tonnes"
  ..@ landings.n  :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0 4792 66428 49659 35282 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "thousands"
  ..@ landings.wt :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0 0.165 0.201 0.258 0.353 0.456 0.533 0.589 0.396 0.998 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "kg"
  ..@ stock       :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1, 1:58, 1, 1, 1, 1] 331622 348810 383681 400595 417042 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr "all"
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "tonnes"
  ..@ stock.n     :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 460518 260189 335426 191663 123254 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "thousands"
  ..@ stock.wt    :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0.038 0.102 0.157 0.242 0.325 0.485 0.719 0.682 0.844 0.918 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "kg"
  ..@ m           :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "NA"
  ..@ mat         :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0 0.5 0.5 1 1 1 1 1 1 1 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "NA"
  ..@ harvest     :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0.0767 0.2276 0.2706 0.3249 0.3698 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "f"
  ..@ harvest.spwn:Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0 0 0 0 0 0 0 0 0 0 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "NA"
  ..@ m.spwn      :Formal class 'FLQuant' [package "FLCore"] with 2 slots
  .. .. ..@ .Data: num [1:10, 1:58, 1, 1, 1, 1] 0 0 0 0 0 0 0 0 0 0 ...
  .. .. .. ..- attr(*, "dimnames")=List of 6
  .. .. .. .. ..$ age   : chr [1:10] "1" "2" "3" "4" ...
  .. .. .. .. ..$ year  : chr [1:58] "1957" "1958" "1959" "1960" ...
  .. .. .. .. ..$ unit  : chr "unique"
  .. .. .. .. ..$ season: chr "all"
  .. .. .. .. ..$ area  : chr "unique"
  .. .. .. .. ..$ iter  : chr "1"
  .. .. ..@ units: chr "NA"
  ..@ name        : chr "Plaice in IV"
  ..@ desc        : chr "Imported from a VPA file. ( W:\\\\IMARES\\\\Data\\\\ICES-WG\\\\Demersale werkgroep WGNSSK\\\\2015\\\\stock\\\\ple-nsea\\\\Final run 2015\\\\index"| __truncated__
  ..@ range       : Named num [1:7] 1 10 10 1957 2014 ...
  .. ..- attr(*, "names")= chr [1:7] "min" "max" "plusgroup" "minyear" ...
}}
\usage{
ple4
}
\description{
Example file: North Sea Place 2015 assessment results
}
\author{
David Miller <david.miller@wur.nl>
}
\keyword{datasets}

