% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rba_icod}
\alias{rba_icod}
\title{Results by age}
\format{\preformatted{'data.frame':	12 obs. of  12 variables:
 $ Age  : int  3 4 5 6 7 8 9 10 11 12 ...
 $ M    : num  0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 0.2 ...
 $ Mat  : num  0.003 0.017 0.0645 0.2895 0.676 ...
 $ WeSt : num  944 1808 2761 3567 4833 ...
 $ F    : num  0.038 0.0929 0.1596 0.2353 0.3148 ...
 $ WeCa : num  1.23 1.76 2.51 3.49 4.75 ...
 $ PF   : num  0 0 0 0 0 0 0 0 0 0 ...
 $ PM   : num  0 0 0 0 0 0 0 0 0 0 ...
 $ Fd   : num  0 0 0 0 0 0 0 0 0 0 ...
 $ WeCad: num  0 0 0 0 0 0 0 0 0 0 ...
 $ Fi   : num  0 0 0 0 0 0 0 0 0 0 ...
 $ WeCai: num  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
rba_icod
}
\description{
Example file: Cod results by age
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

