% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvk_to_converter.R
\name{rbx_to_flstock}
\alias{rbx_to_flstock}
\title{Convert rvk-format to FLStock object}
\usage{
rbx_to_flstock(
  rbx,
  sName = "nn",
  sDesc = "none",
  pf = 0,
  pm = 0,
  aF1 = 5,
  aF2 = 10
)
}
\arguments{
\item{rbx}{A list with two data.frames named rby and rbya.}

\item{sName}{Character vector containing name of stock}

\item{sDesc}{Character vectro containing some description}

\item{pf}{Partial fishing mortality}

\item{pm}{Partial natural mortality}

\item{aF1}{Lower reference age}

\item{aF2}{Upper reference age}
}
\description{
Converts the rby and rbya to FLStock object.
}
\note{
TO DO: check the unit stuff
}
