% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rby_icod}
\alias{rby_icod}
\title{Results by year}
\format{An object of class \code{data.frame} with 61 rows and 5 columns.}
\usage{
rby_icod
}
\description{
Example file: Cod results by year.
}
\author{
Einar Hjorleifsson <einar.hjorleifsson@gmail.com>
}
\keyword{datasets}

