% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rw_lowestoft.R
\name{read_lowestoft}
\alias{read_lowestoft}
\title{read_lowestoft_file}
\usage{
read_lowestoft(file, format = "matrix", value.name = "x", sep = "",
  quiet = TRUE)
}
\arguments{
\item{file}{name of file, normally the index file name}

\item{format}{output format, "matrix","wide","long"}

\item{value.name}{Name of value, only relevant for format="long"}

\item{sep}{the separator, default is ""}

\item{quiet}{boolean, default is TRUE}
}
\description{
The code is inpired by the
 \href{https://github.com/flr/FLCore/blob/master/R/io.VPAsuite.R}{FLCore::readVPAFile} function.
The difference is that it is not dependent on the FLCore, including the S4-methods.
}

