% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_ibya}
\alias{sam_ibya}
\title{sam_ibya}
\usage{
sam_ibya(fit)
}
\arguments{
\item{fit}{A "sam"-object}
}
\value{
A tibble containing the following variables:
\itemize{
  \item year:
  \item age:
  \item cW: Catch weights
  \item dW: Discards weights
  \item lF: Fraction of fish landed
  \item lW: Landed weights
  \item m: Assumed natural mortality (M)
  \item mat: Maturity ogive
  \item oC: Observed catch
  \item oU1: Observed survey index 1
  \item pU2: Observed survey index 2, etc.
  \item pF: Proportional F prior to spawning
  \item pM: Proportional M prior to spawning
  \item sW: Spawning stock weight
}
}
\description{
sam input data as a tibble
}
