% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_process_error}
\alias{sam_process_error}
\title{sam process error}
\usage{
sam_process_error(rbya, plot_it = FALSE, plot_catch = FALSE, plus_group = TRUE)
}
\arguments{
\item{rbya}{XXX}

\item{plot_it}{XXX}

\item{plot_catch}{XXX}

\item{plus_group}{XXX}
}
\description{
XXX
}
