% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sam_process_error.R
\name{sam_process_error}
\alias{sam_process_error}
\title{sam process error}
\usage{
sam_process_error(rbya, plus_group = TRUE)
}
\arguments{
\item{rbya}{XXX}

\item{plus_group}{XXX}
}
\description{
XXX
}

