% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_process_error.R
\name{sam_process_error}
\alias{sam_process_error}
\title{sam process error}
\usage{
sam_process_error(rbya, plus_group = TRUE, plot_it = FALSE)
}
\arguments{
\item{rbya}{XXX}

\item{plus_group}{XXX}

\item{plot_it}{XXX}
}
\description{
XXX
}

