% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam_rbx}
\alias{sam_rbx}
\title{rbx tibbles from object class "sam"}
\usage{
sam_rbx(fit, scale = 1)
}
\arguments{
\item{fit}{A "sam"-object}

\item{scale}{A scaler (default 1)}
}
\value{
A list containing tibbles "rby", "rbya" and "opr"
}
\description{
rbx tibbles from object class "sam"
}
